split fs2 loss module
